refactor: split app session planner helpers
